Add files via upload
The Dataset#1 is composed of the Tweets posted by Top French football clubs (psg, om, ol, fcnantes, losc, asmonaco, asse). 
The text was analyzed using modern NPL libraries and techniques. That includes an analysis of the punctuation and the words influencing the retweets frequency, a sentiment analysis and topic modeling.