Add files via upload
Dataset#3 includes only Tweets from fans that posted with a hashtag related to Olympique de Marseille. 
It is interesting to gather the question marks posted by fans and react quickly to the questions they might have. 
Secondly, Topic modeling for a specific club is particularly relevant, and enables the club to create innovative segmentation strategies.